archive/requirements.txt
RPi.GPIO
smbus2
